+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 0.1.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- no more open browser by default
+  pytest-play is a generic test engine and it could be used for non UI tests too.
+
+  So there is no need to open the browser for non UI tests (eg: API tests)
 
 
 0.1.0 (2017-12-22)